+++ R/TRI-model.R
@@ -1,3 +1,45 @@
+
+triangulate_0 <- function(x, ...) {
+  objlist <- split(x$path, x$path$object_)
+  objlist <- objlist[unique(x$path$object_)]
+  polygon_count <- nrow(dplyr::distinct(x$path[c("object_", "subobject")]))
+  trilist <- vector("list", polygon_count)
+  itri <- 0
+  for (i in seq_along(objlist)) {
+    #obj <- objlist[[i]]
+    #browser()
+    if (length(unique(objlist[[i]]$subobject)) > 1) {
+      subobjlist <- split(objlist[[i]], objlist[[i]]$subobject)
+      subobjlist <- subobjlist[unique(objlist[[i]]$subobject)]
+    } else {
+      subobjlist <- objlist[i]
+    }
+    for (j in seq_along(subobjlist)) {
+      itri <- itri + 1
+      verts <- subobjlist[[j]] %>%
+        dplyr::select(.data$object_, .data$path_) %>%
+        dplyr::inner_join(x$path[c("path_", "object_")], "path_") %>%
+        dplyr::select(.data$path_) %>%
+        dplyr::inner_join(x$path_link_vertex, "path_") %>%
+        dplyr::inner_join(x$vertex, "vertex_")
+      holes <- which(c(0, abs(diff(as.integer(as.factor(verts$path_))))) > 0)
+      if (length(holes) < 1) holes <- 0
+
+      trindex <- decido::earcut(cbind(verts[["x_"]], verts[["y_"]]), holes)
+      trimat <- matrix(trindex, ncol = 3L, byrow = TRUE)
+      trilist[[itri]] <- tibble::tibble(.vx0 = verts$vertex_[trimat[,1L]],
+                                        .vx1 = verts$vertex_[trimat[,2L]],
+                                        .vx2 = verts$vertex_[trimat[,3L]],
+                                        path_ = verts$path_[1L],
+                                        object_ = objlist[[i]]$object_[1L])
+
+    }
+  }
+  dplyr::bind_rows(trilist)
+}
+
+
+
 #' TRI model, triangulations
 #'
 #' TRI creates a constrained triangulation using 'ear-cutting', or 'ear-clipping' of
@@ -21,6 +63,11 @@ TRI.default <- function(x, ...) {
   ## TRI is earcut, so must be PATH based
   TRI(PATH(x), ...)
 }
+#' @export
+TRI.mesh3d <- function(x, ...) {
+  TRI(TRI0(x), ...)
+}
+
 #' @export
 TRI.TRI <- function(x, ...) {
   x
@@ -49,6 +96,11 @@ TRI.TRI0 <- function(x, ...){
 #' @export
 TRI.PATH0 <- function(x, ...) {
   TRI(PATH(x), ...)
+}
+#' @name TRI
+#' @export
+TRI.sfc_GEOMETRYCOLLECTION <- function(x, ...) {
+  TRI(TRI0(x), ...)
 }
 #' @export
 TRI.PATH <- function(x, ...) {