+++ NEWS.md
@@ -1,3 +1,13 @@
+# silicate dev
+
+* New TRI and TRI0 methods for mesh3d. No material properties yet, just one mesh object. 
+
+* Added TRI and TRI0 methods for sfc_GEOMETRYCOLLECTION to catch lists of
+POLYGON triangles. Rudimentary function now for TRI0/TRI (straight conversion), 
+PATH0, ARC. Triangles automatically include `z_ = 0` if not already present.
+
+* New data set `ear_gc` a GEOMETRYCOLLECTION of POLYGON triangles. 
+
 # silicate 0.4.0
 
 * Fixed obscure bug that messed up anglr because of class-dropping for the